Split tmux reconnect chrome
